Move frontend components into separate folder
